// backend/src/types/response/vip/getVipTierRewardResponse.ts
export type GetVipTierRewardResponse = {
  vip_tier_id: number | null;
  point_redeemed_id: number | null;
  reward_key: string | null;
  reward_key_key_display_text: string | null;
  reward_icon: string | null;
  reward_description: string | null;
  is_reward_enabled: string | null;
  status: string | null;
  created_by: number | null;
  updated_by: number | null;
  admin_ref: number | null;
  created_at: Date;
  updated_at: Date | null;
};